pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sudp"
version = "0.4.0"
description = "Simple UDP Protocol"
requires-python = ">=3.8"
authors = [
    {name = "Meir Michanie"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: No Input/Output Interaction",
    "Intended Audience :: System Administrators",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: System :: Networking"
]
dependencies = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.0",
    "pyyaml==6.0.1",
    "websockets==12.0",
    "python-dotenv==1.0.0"
]

[project.scripts]
sudpd = "sudp.server.daemon:main"
sudpc = "sudp.client.daemon:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["sudp*"] 